add holiday support and fix web initializer

+++ attendance/src/main/scala/org/openurp/ws/services/teach/attendance/impl/DataImporter.scala
@@ -36,41 +36,45 @@ import org.openurp.ws.services.teach.attendance.domain.WeekStates
 class DataImporter extends Logging {
 
   var executor: JdbcExecutor = _
+  var appConfig: AppConfig = _
+  var baseDataService: BaseDataService = _
   private val lockName = "tmp_attend_importer_lock"
 
   def importData(date: Date): (Int, Int) = {
-    var rs = WeekStates.build(date)
-    val count = executor.queryForInt("select count(*) from user_tables where table_name='" + lockName.toUpperCase + "'")
     var activityCountSum, detailCount = 0
-    if (0 == count) {
-      logger.info("Attend Importer {}  executing...", toDateStr(date))
-      executor.update("create table " + lockName + "(id number(10))")
-      try {
-        //具体考勤活动数据
-        val activityCntList = executor.query("select count(*) from " + activityTable(date) + " where course_date=?", date)
-        if (activityCntList.isEmpty || activityCntList.head.head.asInstanceOf[Number].intValue == 0) {
-          val sql = "insert into " + activityTable(date) + "(id,semester_id,lesson_id,course_id,course_date,begin_time,end_time,room_id,course_hours,attend_begin_time)" +
-            " select seq_t_attend_activities.nextval,rw.jxrlid,rw.id,rw.kcid,to_date(" + toDateStr(date) + ",'yyyyMMdd')," +
-            " hd.qssj,hd.jssj,hd.jsid,(hd.jsxj-hd.ksxj+1) ks,to_coursetime(to_minutes(hd.qssj)-15) begin_attend_time from jxrw_t rw ,jxhd_t hd" +
-            " where rw.id=hd.jxrwid and hd.zj=? and hd.nf=? and bitand(hd.yxzsz,?)>0 and hd.jsid is not null and hd.zyyy=1"
-          rs._2.foreach { data =>
-            val activityCount = executor.update(sql, rs._1, data._1, data._2)
-            activityCountSum += activityCount
-            logger.info("Importe {} attend activities", activityCount)
+    if (!baseDataService.isHoliday(date)) {
+      var rs = WeekStates.build(date)
+      val count = executor.queryForInt("select count(*) from user_tables where table_name='" + lockName.toUpperCase + "'")
+      if (0 == count) {
+        logger.info("Attend Importer {}  executing...", toDateStr(date))
+        executor.update("create table " + lockName + "(id number(10))")
+        try {
+          //具体考勤活动数据
+          val activityCntList = executor.query("select count(*) from " + activityTable(date) + " where course_date=?", date)
+          if (activityCntList.isEmpty || activityCntList.head.head.asInstanceOf[Number].intValue == 0) {
+            val sql = "insert into " + activityTable(date) + "(id,semester_id,lesson_id,course_id,course_date,begin_time,end_time,room_id,course_hours,attend_begin_time)" +
+              " select seq_t_attend_activities.nextval,rw.jxrlid,rw.id,rw.kcid,to_date(" + toDateStr(date) + ",'yyyyMMdd')," +
+              " hd.qssj,hd.jssj,hd.jsid,(hd.jsxj-hd.ksxj+1) ks,to_coursetime(to_minutes(hd.qssj)-" + appConfig.lateMax + ") begin_attend_time from jxrw_t rw ,jxhd_t hd" +
+              " where rw.id=hd.jxrwid and hd.zj=? and hd.nf=? and bitand(hd.yxzsz,?)>0 and hd.jsid is not null and hd.zyyy=1"
+            rs._2.foreach { data =>
+              val activityCount = executor.update(sql, rs._1, data._1, data._2)
+              activityCountSum += activityCount
+              logger.info("Importe {} attend activities", activityCount)
+            }
           }
+          //上课名单数据
+          val detailCntList = executor.query("select count(*) from " + detailTable(date) + " d," + activityTable(date) + " aa where d.activity_id=aa.id " +
+            " and aa.course_date =?", date)
+          if (detailCntList.isEmpty || detailCntList.head.head.asInstanceOf[Number].intValue == 0) {
+            val detailSql = "insert into " + detailTable(date) + "(id,std_id,activity_id,attend_type_id)" +
+              " select seq_t_attend_details.nextval,jxb.xsid,aa.id," + AttendType.Absenteeism + " from " + activityTable(date) + " aa,jxbxs_t jxb" +
+              " where aa.course_date = ? and aa.lesson_id=jxb.jxrwid"
+            detailCount = executor.update(detailSql, date)
+            logger.info("Importe {} attend details.", detailCount)
+          }
+        } finally {
+          executor.update("drop table " + lockName)
         }
-        //上课名单数据
-        val detailCntList = executor.query("select count(*) from " + detailTable(date) + " d," + activityTable(date) + " aa where d.activity_id=aa.id " +
-          " and aa.course_date =?", date)
-        if (detailCntList.isEmpty || detailCntList.head.head.asInstanceOf[Number].intValue == 0) {
-          val detailSql = "insert into " + detailTable(date) + "(id,std_id,activity_id,attend_type_id)" +
-            " select seq_t_attend_details.nextval,jxb.xsid,aa.id," + AttendType.Absenteeism + " from " + activityTable(date) + " aa,jxbxs_t jxb" +
-            " where aa.course_date = ? and aa.lesson_id=jxb.jxrwid"
-          detailCount = executor.update(detailSql, date)
-          logger.info("Importe {} attend details.", detailCount)
-        }
-      } finally {
-        executor.update("drop table " + lockName)
       }
     }
     (activityCountSum, detailCount)
